10/16/21 Day Session
Did a lot today. Added sorting to the selected cards, created a 'Clear Selection' button, refactored the S10Controller2 script to pull the card GameObjects directly rather than use an abstracted card class. This shortened the code a lot, then I converted some for loops to foreach blocks for simplicity. Added a submit button, and next step is to make it functional.