10/26/21 Night Session
Wrote a function to compare the two hands, using the characterize function to see what the two hands are. There is a bug where if they are the same type it should compare their first card, but the code isn't reaching that block for some reason. Next step is to fix that.