11/15/21 Night Session
Moved scripts not directly dealing with clicking a card to a centralized script instead, creating all player objects, and wrote a function to select and play a card for a computer player. Next step is to write a function to automatically rotate through all computers and come back to me.