11/28/21 Night Session
Created change turn function to change color of current player. Also realized the Destroy function destroys the object itself in the editor when the script runs in the editor. The script ran recursively non-stop (a bug) in the editor (which I didn't know was possible), thus deleting two thirds of my cards. So I recreated those and changed Destroy to inactivate. Next is to work on StealTurn button for human.