11/29/21 Night Session
Added StealTurn method and attached to button for human to steal. Figured out how to stop Rotate function to allow me to jump in, then restart it when I'm done. I can play through most of a game now, just need to fix a few bugs and add conditions to find rankings at the end and then finish the game.