12/21/21 Night Session
Finished the How To Play page, added buttons to open and close it, made the game 16:9 aspect ratio, cleaned up a few bugs. A user can now learn the game and play through multiple games! I have a list of nice to have items including a settings page, animation of the cards, and making the computer players stronger, which I hope to get to soon.